backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
SQLAlchemy==2.0.*
alembic==1.13.*
pydantic==2.*
email-validator==2.*
PyJWT==2.9.*
argon2-cffi==23.*
python-multipart==0.0.9
slowapi==0.1.*
psycopg2-binary==2.9.*
structlog==24.*
python-dotenv==1.0.*